+++ examples/IRDispatcherDemo/IRCommandDispatcher.h
@@ -82,7 +82,7 @@ public:
     void checkAndCallCommand(bool aCallAlsoBlockingCommands);
 
     void printIRCommandString(Print *aSerial);
-    void setRequestToStopReceived();
+    void setRequestToStopReceived(bool aRequestToStopReceived = true);
 
     uint8_t currentBlockingCommandCalled = COMMAND_INVALID; // The code for the current called command
     bool executingBlockingCommand = false;              // Lock for recursive calls of regular commands